Reaping test: remove no sleep assumption
The reaping test assumes that all the `sleep` processes on the system
are created by us, which is an incorrect assumption, and creates
problems now that we run in a more "complete" system in CI (i.e. not in
a minimal container).

This solves the issue by not assuming that random sleep processes are
related to us: instead, we look at whether these processes are in our
process group.

This also includes a little clean up of the reaping stage 1.