Add post-processing command
Post processing commands can be specified on the command line. These
commands are executed after the main command exits and can optionally be
passed the exit code from the last process that tini ran.  The exit code
of the post processing command is the new exit code of tini. Multiple
post processing commands can be chained together, with the output of
each one getting the exit code of the previous one. Post processing
commands inherit the signal mask of the parent tini process (unlike the
primary child).