Support bpf syscalls by passing seccomp notify fd in tini
This adds optional support for running bpf programs in Titus
under titus-executor using the new kernel feature that allows
other processes to get notifications of syscall events.

This patch adds a new functionality to tini to support passing
a seccomp-notify-fd to another agent that can do syscalls on
its behalf.